+++ HISTORY.md
@@ -4,6 +4,9 @@ Locator Lang Change History
 @NEXT@
 ------------------
 
+0.2.0 (2014-03-06)
+------------------
+
 * PR #6: supporting requiredLangs of langs to fallback lang bundles and lang entries when needed
 
 0.1.2 (2014-02-13)